Extract HIE.Bios.Process module
